refac: Moved code that is resposible for state of button
Moved code that is handling button click from App to buttons component
Cleaned the code of unnecessary divs
