+++ src/App.js
@@ -13,72 +13,9 @@ function App() {
     stone: false,
   });
 
-  const resourceHandler = (i) => {
-    if (i == "0") {
-      setVisibility({
-        ...visibility,
-        bank: true,
-      });
-      if (visibility.bank) {
-        setVisibility({
-          ...visibility,
-          bank: false,
-        });
-      }
-    }
-    if (i == "1") {
-      setVisibility({
-        ...visibility,
-        goldExport: true,
-      });
-      if (visibility.goldExport) {
-        setVisibility({
-          ...visibility,
-          goldExport: false,
-        });
-      }
-    }
-    if (i == "2") {
-      setVisibility({
-        ...visibility,
-        goldOre: true,
-      });
-      if (visibility.goldOre) {
-        setVisibility({
-          ...visibility,
-          goldOre: false,
-        });
-      }
-    }
-    if (i == "3") {
-      setVisibility({
-        ...visibility,
-        ironOre: true,
-      });
-      if (visibility.ironOre) {
-        setVisibility({
-          ...visibility,
-          ironOre: false,
-        });
-      }
-    }
-    if (i == "4") {
-      setVisibility({
-        ...visibility,
-        stone: true,
-      });
-      if (visibility.stone) {
-        setVisibility({
-          ...visibility,
-          stone: false,
-        });
-      }
-    }
-  };
-
   return (
     <div className="row mx-5 my-5">
-      <Menu resourceHandler={resourceHandler} />
+      <Menu visibility={visibility} setVisibility={setVisibility} />
       <div className="col position-relative">
         <Map />
         <LoopResourceMarks visibility={visibility} />
