+++ src/components/Menu/Menu.js
@@ -1,13 +1,10 @@
 import classes from "./Menu.module.css"
-
 import ResourceList from "./ResourceList/ResourceList";
 
-import { RESOURCES_DATA } from "../../resourceData";
-
-function Menu({resourceHandler}) {
+function Menu({visibility, setVisibility}) {
   return (
     <div className={"col col-lg-2 " + classes.menu}>
-        <ResourceList resourceHandler={resourceHandler} resources={RESOURCES_DATA}/>
+        <ResourceList visibility={visibility} setVisibility={setVisibility}/>
     </div>
   );
 }
