--- src/hideElements.js
@@ -1,3 +1,0 @@
-function hideElements() {
-    document.getElementById('icon').hidden = true
-}
